Streamlit_Bank_Loan_Prediction/requirements.txt
scikit-learn==1.2.2 
numpy
pandas
matplotlib
streamlit==0.76.0
Pillow
seaborn